\name{theNewScheme-package}
\alias{theNewScheme-package}
\alias{theNewScheme}
\docType{package}
\title{
\packageTitle{theNewScheme}
}
\description{
\packageDescription{theNewScheme}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{theNewScheme}
\packageIndices{theNewScheme}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{theNewScheme}

Maintainer: \packageMaintainer{theNewScheme}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
